server/requirements.txt
django==1.8
djangorestframework
django-choices
django-rest-swagger
django-rest-auth
django-allauth
django-crispy-forms
